put together a ddtest for lwb fp
unfortunately doesn't work yet due to loader BS
